--- src/main/java/org/cloudfoundry/autosleep/frontend/servicebroker/configuration/BrokerConfiguration.java
@@ -1,15 +1,0 @@
-package org.cloudfoundry.autosleep.frontend.servicebroker.configuration;
-
-import org.cloudfoundry.community.servicebroker.model.BrokerApiVersion;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class BrokerConfiguration {
-
-    @Bean
-    public BrokerApiVersion brokerApiVersion() {
-        return new BrokerApiVersion();
-    }
-
-}
