--- src/main/java/org/cloudfoundry/autosleep/frontend/servicebroker/service/InvalidParameterException.java
@@ -1,15 +1,0 @@
-package org.cloudfoundry.autosleep.frontend.servicebroker.service;
-
-import lombok.Getter;
-import org.springframework.http.converter.HttpMessageNotReadableException;
-
-@Getter
-public class InvalidParameterException extends HttpMessageNotReadableException {
-
-    private String parameterName;
-
-    public InvalidParameterException(String parameterName, String error) {
-        super("'" + parameterName + "': " + error);
-        this.parameterName = parameterName;
-    }
-}
