--- src/main/java/org/cloudfoundry/autosleep/frontend/servicebroker/service/parameters/ParameterReader.java
@@ -1,12 +1,0 @@
-package org.cloudfoundry.autosleep.frontend.servicebroker.service.parameters;
-
-import org.cloudfoundry.autosleep.frontend.servicebroker.service.InvalidParameterException;
-
-
-public interface ParameterReader<T> {
-
-    String getParameterName();
-
-    T readParameter(Object parameter, boolean withDefault) throws InvalidParameterException;
-
-}
