--- src/main/java/org/cloudfoundry/autosleep/frontend/webui/configuration/StaticResourceConfiguration.java
@@ -1,17 +1,0 @@
-package org.cloudfoundry.autosleep.frontend.webui.configuration;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
-
-@Configuration
-public class StaticResourceConfiguration extends WebMvcConfigurerAdapter {
-    private static final String[] CLASSPATH_RESOURCE_LOCATIONS = {
-            "classpath:/META-INF/resources/", "classpath:/resources/",
-            "classpath:/static/", "classpath:/public/"};
-
-    @Override
-    public void addResourceHandlers(ResourceHandlerRegistry registry) {
-        registry.addResourceHandler("/**").addResourceLocations(CLASSPATH_RESOURCE_LOCATIONS);
-    }
-}
