+++ src/main/java/org/cloudfoundry/autosleep/ui/servicebroker/service/parameters/ParameterReader.java
@@ -1,0 +1,12 @@
+package org.cloudfoundry.autosleep.ui.servicebroker.service.parameters;
+
+import org.cloudfoundry.autosleep.ui.servicebroker.service.InvalidParameterException;
+
+
+public interface ParameterReader<T> {
+
+    String getParameterName();
+
+    T readParameter(Object parameter, boolean withDefault) throws InvalidParameterException;
+
+}
